requirements.txt
pandas 
fpdf
yagmail
openpyxl
python-dotenv